refactor: 날짜 계산 유틸함수 분리
