+++ front/src/components/Calendar/index.tsx
@@ -1,18 +1,8 @@
 import { useState } from 'react';
 import dayjs from 'dayjs';
-import { MonthYear } from '@typings/domain';
+import { getMonthYearDetails } from '@utils/index';
 import DateBox from './DateBox';
 import { DateGrid } from './styles';
-
-const getMonthYearDetails = (initialDate: dayjs.Dayjs): MonthYear => {
-  const month = initialDate.format('MM');
-  const year = initialDate.format('YYYY');
-  const startDate = dayjs(`${year}${month}01`);
-  const firstDOW = Number(startDate.format('d'));
-  const lastDate = Number(startDate.clone().endOf('month').format('DD'));
-
-  return { startDate, firstDOW, lastDate, month, year };
-};
 
 const Calendar = () => {
   const currentDate = dayjs();
