+++ front/src/utils/index.ts
@@ -1,0 +1,14 @@
+import dayjs from 'dayjs';
+import { MonthYear } from '@typings/domain';
+
+const getMonthYearDetails = (initialDate: dayjs.Dayjs): MonthYear => {
+  const month = initialDate.format('MM');
+  const year = initialDate.format('YYYY');
+  const startDate = dayjs(`${year}${month}01`);
+  const firstDOW = Number(startDate.format('d'));
+  const lastDate = Number(startDate.clone().endOf('month').format('DD'));
+
+  return { startDate, firstDOW, lastDate, month, year };
+};
+
+export { getMonthYearDetails };
